+++ app/src/main/java/com/revature/findlawyer/MainActivity.kt
@@ -28,8 +28,8 @@ class MainActivity : ComponentActivity() {
                     //MainOverllScreen ()
                     //MainLawyer()
                     //MainUser()
-                    //LawyerLogIn()
-                    UserLogIn()
+                    LawyerLogIn()
+                    //UserLogIn()
                 }
             }
         }
