+++ app/src/main/java/com/revature/findlawyer/ui/MainUserLawyer.kt
@@ -3,7 +3,7 @@ package com.revature.findlawyer.ui
 import androidx.compose.foundation.Image
 import androidx.compose.foundation.layout.*
 import androidx.compose.material.Button
-import androidx.compose.material.Scaffold
+import androidx.compose.material.MaterialTheme
 import androidx.compose.material.Surface
 import androidx.compose.material.Text
 import androidx.compose.runtime.Composable
@@ -19,6 +19,51 @@ import androidx.compose.ui.unit.dp
 import androidx.compose.ui.unit.sp
 import com.revature.findlawyer.R
 import com.revature.findlawyer.ui.theme.Cormorantgaramond
+
+
+////////////////////////////////////  MAIN LAWYER  /////////////////////////////////////////
+
+@Preview
+@Composable
+fun previewMainLawyer(){
+    MainLawyer()
+}
+
+@Composable
+fun MainLawyer (){
+
+    Surface(modifier = Modifier.fillMaxSize()/*.border(2.dp,color= Color.Red)*/,color = MaterialTheme.colors.background) {
+        //Image(painter = painterResource(id = R.drawable.ic_launcher_background), contentDescription = "",modifier = Modifier.size(400.dp,400.dp).offset(y=200.dp))
+        Image(painter = painterResource(id = R.drawable.citypersp), contentDescription = "",contentScale = ContentScale.FillBounds,alpha = .3f)
+        //Image(painter = painterResource(id = R.drawable.loginimg), contentDescription = "",contentScale = ContentScale.FillBounds,alpha = .4f)
+        Logo()
+        //LogoMain()
+
+
+
+        Column(modifier = Modifier.fillMaxSize()/*.border(2.dp,color= Color.Red)*/.padding(4.dp),verticalArrangement = Arrangement.Bottom,horizontalAlignment = Alignment.CenterHorizontally) {
+
+            Text(text = "LOG IN OR REGISTER TO CONNECT WITH CLIENTS",textAlign = TextAlign.Center, modifier = Modifier.padding(2.dp).offset(y=-55.dp),color= Color.Black,fontFamily = Cormorantgaramond,fontWeight = FontWeight.Bold,fontSize = 17.sp )
+
+            Button(onClick = { /*TODO*/ },modifier= Modifier.padding(4.dp).width(330.dp).offset(y=-50.dp)/*.border(1.dp,color= Color.Black)*/) {
+                Text(text = "Log In", modifier = Modifier.padding(2.dp),color= Color.White)
+
+            }
+            Button(onClick = { /*TODO*/ },modifier= Modifier.padding(4.dp).width(330.dp).offset(y=-50.dp)/*.border(1.dp,color= Color.Black)*/) {
+                Text(text = "Register", modifier = Modifier.padding(2.dp),color= Color.White)
+
+            }
+
+        }
+
+
+    }
+
+}
+
+
+////////////////////////////////////  MAIN USER  /////////////////////////////////////////
+
 
 @Preview
 @Composable
